RC v. 1.3.1 (#273)
## Major changes
- Documentation was improved with three new vignettes about GUI usage, command line usage and output structure.

## Bug fixes
- Restore checking environmental variable in `init_python()` (this urgent bug fixing was the reason of the urgent release of the current version).
- Fix building offline footprint.
- Fix the Dockerfile.

## Changes in default argument values
- Accept both `sen2r(..., rm_safe = "yes"`) and `"all"`.
- In the GUI, the maximum allowed cloud cover was set to 100% accordingly to the `sen2r()` defaults.

## Minor changes
- Add check on `max_mask` - `mask_type` coherence (if `max_mask < 100` and `mask_type` is not specified, a warning is returned).
- Suppress some unuseful warnings.
- Switch examples on 2019 dates (previously examples were based on 2017 images, which were partially moved on LTA).
- Do not print progress bars in logs.

+++ R/safe_getMetadata.R
@@ -91,8 +91,8 @@
 #' @examples
 #' # Define product name
 #' s2_examplenames <- c(
-#'   "S2A_MSIL1C_20170703T101021_N0205_R022_T32TNS_20170703T101041.SAFE",
-#'   "S2A_MSIL2A_20170703T101021_N0205_R022_T32TNR_20170703T101041.SAFE"
+#'   "S2A_MSIL1C_20190723T101031_N0208_R022_T32TNS_20190723T121220.SAFE",
+#'   "S2A_MSIL1C_20190723T101031_N0208_R022_T32TNR_20190723T121220.SAFE"
 #' )
 #'
 #' # Return the information retrievable from the file names (files are not scanned)
@@ -114,9 +114,11 @@
 #' 
 #' \dontrun{
 #' # Download a sample SAFE archive (this can take a while)
-#' s2_exampleurl <- paste0("https://scihub.copernicus.eu/apihub/odata/v1/",
-#'   "Products(\'5f590bcb-ee55-4a20-8e75-bde99f5b93d4\')/$value")
-#' names(s2_exampleurl) <- "S2A_MSIL1C_20170703T101021_N0205_R022_T32TNS_20170703T101041.SAFE"
+#' s2_exampleurl <- c(
+#'   "S2A_MSIL1C_20190723T101031_N0208_R022_T32TNS_20190723T121220.SAFE" =
+#'     paste0("https://scihub.copernicus.eu/apihub/odata/v1/",
+#'            "Products('19bbde60-992b-423d-8dea-a5e0ac7715fc')/$value")
+#' )
 #' s2_download(s2_exampleurl, outdir=tempdir())
 #' s2_examplepath <- file.path(tempdir(), names(s2_exampleurl))
 #'
@@ -577,18 +579,20 @@ safe_isvalid <- function(s2, allow_oldnames = FALSE, check_file = TRUE) {
       }
       
       # check jp2 files exist
-      jp2_listall <- list.files(s2_path, "\\.jp2$", recursive=TRUE, full.names=FALSE)
-      if (length(jp2_listall) == 0) {
-        s2_validname <- FALSE # missing data
-        if (action == "getmetadata") {
-          print_message(
-            type=message_type,
-            "This product (",s2_name,") does not contain raster data, ",
-            "and should be deleted from directory \"",dirname(s2_path),"\"; ",
-            "otherwise, errors can occur during data processing."
-          )
-        } else if (action == "rm_invalid" & s2_exists) {
-          unlink(s2_path, recursive=TRUE)
+      if (s2_validname == TRUE) {
+        jp2_listall <- list.files(s2_path, "\\.jp2$", recursive=TRUE, full.names=FALSE)
+        if (length(jp2_listall) == 0) {
+          s2_validname <- FALSE # missing data
+          if (action == "getmetadata") {
+            print_message(
+              type=message_type,
+              "This product (",s2_name,") does not contain raster data, ",
+              "and should be deleted from directory \"",dirname(s2_path),"\"; ",
+              "otherwise, errors can occur during data processing."
+            )
+          } else if (action == "rm_invalid" & s2_exists) {
+            unlink(s2_path, recursive=TRUE)
+          }
         }
       }
       
@@ -778,9 +782,9 @@ safe_isvalid <- function(s2, allow_oldnames = FALSE, check_file = TRUE) {
             matrix(sel_footprint_raw0, ncol = 2, byrow = TRUE)[,2:1],
             1, paste, collapse = " "
           )
-          metadata[[i]][["footprint"]] <- paste(
-            paste0("POLYGON((",sel_footprint_raw1,"))"),  
-            collapse = ", "
+          metadata[[i]][["footprint"]] <-  paste0(
+            "POLYGON((",
+            paste(sel_footprint_raw1,  collapse = ", "),"))"
           )
         }
       }
